requirements.txt
altair==6.0.0
asttokens==3.0.1
attrs==25.4.0
babel==2.17.0
backgroundremover==0.3.5
blinker==1.9.0
branca==0.8.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
comm==0.2.3
commandlines==0.4.1
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
folium==0.20.0
folium-arrow-icon==0.1.3
fonttools==4.61.0
fsspec==2025.12.0
future==1.0.0
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hsh==1.1.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kaleido==1.2.0
kiwisolver==1.4.9
kneed==0.8.5
lazy_loader==0.4
llvmlite==0.45.1
logistro==2.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
narwhals==2.13.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
numba==0.62.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
orjson==3.11.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
pillow_heif==1.1.1
pip3-autoremove==2.0.1
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
proglog==0.1.12
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydeck==0.9.1
PyDocX==0.9.10
Pygments==2.19.2
PyMatting==1.1.14
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.2
PySocks==1.7.1
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
scikit-image==0.25.2
scipy==1.16.3
selenium==4.39.0
setuptools==80.9.0
shapely==2.1.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.10.16
tk==0.1.0
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.4
wheel==0.45.1
wsproto==1.3.2
xarray==2025.12.0
xyzservices==2025.11.0
